--- blueair/__init__.py
@@ -1,7 +1,0 @@
-"""Unofficial Blueair Python client."""
-
-from .blueair import BlueAir
-
-__version__ = "1.0.0"
-
-__all__ = ["BlueAir"]
