+++ custom_components/blueair/entity.py
