+++ custom_components/blueair/fan.py
