+++ custom_components/blueair/translations/en.json
